+++ packages/core/src/components/Table/Body/Body.tsx
@@ -1,5 +1,7 @@
-import React, { useContext } from 'react';
+import React, { useContext, useMemo } from 'react';
 import { TableComponentsCommonPropsContext } from '../context';
+import { getGridTemplateColumns } from '../helpers';
+import Minimap from '../Minimap';
 import { TBody } from './Body.styled';
 import GroupedRow from './GroupedRow';
 import Row from './Row';
@@ -7,8 +9,13 @@ import { NoResultCell, NoResultRow } from './Row/Row.styled';
 import { Props } from './types';
 
 const Body: React.FC<Props> = React.memo(props => {
-    const { data, groupBy, rowIdentifier, showRowWithCardStyle, noResultRow } = useContext(TableComponentsCommonPropsContext),
-        { selectedRowIds, onRowSelection, onGroupedRowSelection, setUniqueIds, ...restProps } = props;
+    const { data, groupBy, rowIdentifier, showRowWithCardStyle, noResultRow, tableRef, withMinimap, columns, size } = useContext(
+            TableComponentsCommonPropsContext
+        ),
+        { selectedRowIds, onRowSelection, onGroupedRowSelection, setUniqueIds, ...restProps } = props,
+        /* since minimap is positioned sticky with respect to the tbody, tbody should have full table width otherwise minimap positioning fails */
+        tableVisibleWidth = tableRef.current?.clientWidth ?? 0,
+        minimapDimensionDeps = useMemo(() => [columns], [columns]);
 
     return (
         <TBody>
@@ -16,8 +23,14 @@ const Body: React.FC<Props> = React.memo(props => {
                 (noResultRow ? (
                     noResultRow
                 ) : (
-                    <NoResultRow showRowWithCardStyle={showRowWithCardStyle}>
-                        <NoResultCell>No result</NoResultCell>
+                    <NoResultRow
+                        showRowWithCardStyle={showRowWithCardStyle}
+                        gridTemplateColumns={getGridTemplateColumns(columns)}
+                        withMinimap={withMinimap}
+                    >
+                        <NoResultCell width={tableVisibleWidth} tableSize={size}>
+                            No result
+                        </NoResultCell>
                     </NoResultRow>
                 ))}
             {data.map((row, index) => {
@@ -36,6 +49,7 @@ const Body: React.FC<Props> = React.memo(props => {
                     <Row id={identifier} key={identifier} data={row} {...{ ...restProps, selectedRowIds, onRowSelection }} />
                 );
             })}
+            {withMinimap && <Minimap tableRef={tableRef} minimapDimensionDeps={minimapDimensionDeps} />}
         </TBody>
     );
 });