+++ packages/core/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.46.9](https://github.com/medly/medly-components/compare/@medly-components/core@1.46.8...@medly-components/core@1.46.9) (2021-02-25)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.8 (2021-02-25)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.7 (2021-02-24)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.6 (2021-02-24)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.5 (2021-02-24)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.4 (2021-02-24)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.3 (2021-02-22)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.2 (2021-02-22)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+## 1.46.1 (2021-02-22)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
 # [1.46.0](https://github.com/medly/medly-components/compare/@medly-components/core@1.45.0...@medly-components/core@1.46.0) (2021-02-17)